Fill rules and add new rules. (#1)
